requirements.txt
faiss-cpu>=1.7.4
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
